Make CoordinateSpace::logical_size return NonZeroUsize

+++ src/test/box_coordinate_space.rs
@@ -4,26 +4,26 @@ use crate::interface::point::CoordinateSpace;
 #[test]
 fn test_size() {
     // One-dimensional
-    assert_eq!(BoxCoordinateSpace::<1>::new_checked([1]).logical_size(), 1);
-    assert_eq!(BoxCoordinateSpace::<1>::new_checked([2]).logical_size(), 2);
-    assert_eq!(BoxCoordinateSpace::<1>::new_checked([3]).logical_size(), 3);
-    assert_eq!(BoxCoordinateSpace::<1>::new_checked([9]).logical_size(), 9);
-    assert_eq!(BoxCoordinateSpace::<1>::new_checked([10]).logical_size(), 10);
+    assert_eq!(usize::from(BoxCoordinateSpace::<1>::new_checked([1]).logical_size()), 1);
+    assert_eq!(usize::from(BoxCoordinateSpace::<1>::new_checked([2]).logical_size()), 2);
+    assert_eq!(usize::from(BoxCoordinateSpace::<1>::new_checked([3]).logical_size()), 3);
+    assert_eq!(usize::from(BoxCoordinateSpace::<1>::new_checked([9]).logical_size()), 9);
+    assert_eq!(usize::from(BoxCoordinateSpace::<1>::new_checked([10]).logical_size()), 10);
 
-    assert_eq!(BoxCoordinateSpace::<2>::new_checked([1, 1]).logical_size(), 1);
-    assert_eq!(BoxCoordinateSpace::<2>::new_checked([2, 1]).logical_size(), 2);
-    assert_eq!(BoxCoordinateSpace::<2>::new_checked([1, 2]).logical_size(), 2);
-    assert_eq!(BoxCoordinateSpace::<2>::new_checked([2, 2]).logical_size(), 4);
-    assert_eq!(BoxCoordinateSpace::<2>::new_checked([9, 9]).logical_size(), 81);
-    assert_eq!(BoxCoordinateSpace::<2>::new_checked([61, 73]).logical_size(), 4453);
+    assert_eq!(usize::from(BoxCoordinateSpace::<2>::new_checked([1, 1]).logical_size()), 1);
+    assert_eq!(usize::from(BoxCoordinateSpace::<2>::new_checked([2, 1]).logical_size()), 2);
+    assert_eq!(usize::from(BoxCoordinateSpace::<2>::new_checked([1, 2]).logical_size()), 2);
+    assert_eq!(usize::from(BoxCoordinateSpace::<2>::new_checked([2, 2]).logical_size()), 4);
+    assert_eq!(usize::from(BoxCoordinateSpace::<2>::new_checked([9, 9]).logical_size()), 81);
+    assert_eq!(usize::from(BoxCoordinateSpace::<2>::new_checked([61, 73]).logical_size()), 4453);
 
-    assert_eq!(BoxCoordinateSpace::<3>::new_checked([1, 1, 1]).logical_size(), 1);
-    assert_eq!(BoxCoordinateSpace::<3>::new_checked([2, 1, 1]).logical_size(), 2);
-    assert_eq!(BoxCoordinateSpace::<3>::new_checked([1, 2, 1]).logical_size(), 2);
-    assert_eq!(BoxCoordinateSpace::<3>::new_checked([2, 2, 1]).logical_size(), 4);
-    assert_eq!(BoxCoordinateSpace::<3>::new_checked([2, 2, 2]).logical_size(), 8);
-    assert_eq!(BoxCoordinateSpace::<3>::new_checked([9, 9, 9]).logical_size(), 729);
-    assert_eq!(BoxCoordinateSpace::<3>::new_checked([61, 73, 12]).logical_size(), 53436);
+    assert_eq!(usize::from(BoxCoordinateSpace::<3>::new_checked([1, 1, 1]).logical_size()), 1);
+    assert_eq!(usize::from(BoxCoordinateSpace::<3>::new_checked([2, 1, 1]).logical_size()), 2);
+    assert_eq!(usize::from(BoxCoordinateSpace::<3>::new_checked([1, 2, 1]).logical_size()), 2);
+    assert_eq!(usize::from(BoxCoordinateSpace::<3>::new_checked([2, 2, 1]).logical_size()), 4);
+    assert_eq!(usize::from(BoxCoordinateSpace::<3>::new_checked([2, 2, 2]).logical_size()), 8);
+    assert_eq!(usize::from(BoxCoordinateSpace::<3>::new_checked([9, 9, 9]).logical_size()), 729);
+    assert_eq!(usize::from(BoxCoordinateSpace::<3>::new_checked([61, 73, 12]).logical_size()), 53436);
 }
 
 #[test]
